requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
